Separate single sample test from multi samples test Also not common use of same sample between tests
